+++ src/app/mod.rs
@@ -732,3 +732,6 @@ impl Ender {
         Ok(SigType::Interrupt)
     }
 }
+
+#[cfg(test)]
+mod tests;